Fixed various warnings in Visual Studio

+++ engine/resources/codecs/atlas_codec.cpp
@@ -65,7 +65,7 @@ atlas_ptr atlas_codec::decode(resource_manager &parent, gfx::device &device, atl
 
         atlas_region region(region_rect.left, region_rect.top, region_rect.right, region_rect.bottom, size);
 
-        names.insert({region_entry.first, regions.size()});
+        names.insert({region_entry.first, static_cast<int>(regions.size())});
         regions.push_back(region);
     }
